Refactor CancellableRunnable
